requirements.txt
Django==5.0.0
python-dotenv==1.0.0
google-generativeai==0.8.3
groq==0.11.0